+++ tests/unit/test_film.py
@@ -55,9 +55,10 @@ def test_read_highlight_events(events: list[highlight_events.HighlightEvent]):
         ("38_63391382-6b88-43f9-9ff4-6313404c419c.gzip", 2535445291321133, "aCurtis X89", 47053, 203, 203, 149, 85),
         ("39_73607aca-3a23-4733-9e84-b7c2ea679ce2.gzip", 2712212454167317, "I1 DUCE", 45858, 130, 130, 98, 59),
         ("40_fbf6b19e-2b1d-419e-8d64-756f33fda990.gzip", 2535445291321133, "aCurtis X89", 57801, 96, 96, 0, 34),
+        ("41_e4309281-7d3c-420c-a9a4-ca2798e56692.gzip", 2535445291321133, "aCurtis X89", 37907, 116, 116, 86, 47),
         # fmt: on
     ],
-    ids=[30, 31, 33, 34, 35, 36, 37, 38, 39, 40],
+    ids=[30, 31, 33, 34, 35, 36, 37, 38, 39, 40, 41],
 )
 def test_read_film_major_versions(
     file_name: str, xuid, gamertag, time_ms, kills, deaths, modes, medals